[step3/로또(자동)] refactor: 테스트 클래스 분리
